+++ commands/dns/dns.go
@@ -4,6 +4,7 @@ import (
 	"github.com/ionos-cloud/ionosctl/v6/commands/dns/dnssec"
 	"github.com/ionos-cloud/ionosctl/v6/commands/dns/quota"
 	"github.com/ionos-cloud/ionosctl/v6/commands/dns/record"
+	reverse_record "github.com/ionos-cloud/ionosctl/v6/commands/dns/reverse-record"
 	"github.com/ionos-cloud/ionosctl/v6/commands/dns/zone"
 	"github.com/ionos-cloud/ionosctl/v6/internal/core"
 	"github.com/spf13/cobra"
@@ -19,6 +20,7 @@ func DNSCommand() *core.Command {
 	}
 	cmd.AddCommand(zone.ZoneCommand())
 	cmd.AddCommand(record.RecordCommand())
+	cmd.AddCommand(reverse_record.Root())
 	cmd.AddCommand(quota.Root())
 	cmd.AddCommand(dnssec.Root())