+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Added
 - Added support for DNS resources:
-  - `ionosctl dns keys` commands which allows you to enable/disable DNSSEC and manage DNSKEY records
-  - `ionosctl dns quota` commands which allows you to get the DNS quota for your account
+  - `ionosctl dns keys` commands which allows you to enable/disable DNSSEC and manage DNSKEY records.
+  - `ionosctl dns quota` commands which allows you to get the DNS quota for your account.
+  - `ionosctl dns reverse-record` commands which allows you to manage reverse DNS records.
 
 ### Fixed
 - Fixed the column path mapping for 'server' resource to display the actual server's type ('CUBE'/'ENTERPRISE'),